fix(security,quality): resolve Tier 1 CodeQL errors/security alerts
Highest-severity CodeQL code-scanning alerts (errors + security):

- js/xss + js/xss-through-exception (4): harden operator-shell
  appendChildren so only real DOM Nodes are appended raw; every other
  value (incl. exception/user text) is forced through a text node and
  can never be reinterpreted as HTML. Rebuilt the committed bundle.
- cs/loss-of-precision (3): cast to double before the multiplication
  feeding DateTime.AddSeconds/AddMinutes so the product can't overflow
  as int (OutboundDispatcher backoff x2, PdmsDemoSeeder seed offset).
- cs/equality-on-floats (2): script-return truthiness now uses
  `d < 0 || d > 0` — JS-accurate (0 and NaN both falsy) and free of
  float equality operators (JavascriptRouteFilter, ExternalScriptRouteFilter).
- cs/virtual-call-in-constructor (1): ModuleWebApplicationFactory builds
  its Postgres Testcontainer lazily instead of in the base constructor,
  so the abstract ModuleSlug is only read once the derived instance is
  fully constructed.

The 4 cs/unsynchronized-static-access alerts are ImmutableDictionary
<,>.Empty (a thread-safe immutable BCL singleton) — dismissed
separately as false positives rather than contorting correct code.

Build clean; SmartConnect 250/250, HIE 7/7, arch 15/15 green.

+++ src/backend/HIE/Dialysis.HIE.Outbound/Dispatch/OutboundDispatcher.cs
@@ -90,7 +90,7 @@ public sealed class OutboundDispatcher(
             }
             else
             {
-                var retryAt = timeProvider.GetUtcNow().UtcDateTime.AddSeconds(_options.BackoffSeconds * bundle.Attempts);
+                var retryAt = timeProvider.GetUtcNow().UtcDateTime.AddSeconds((double)_options.BackoffSeconds * bundle.Attempts);
                 bundle.MarkAttemptFailed(result.FailureReason ?? $"HTTP {result.StatusCode}", retryAt, _options.MaxAttempts);
                 if (bundle.Status == OutboundBundleStatus.Failed)
                     await EmitFailedAsync(bundle, cancellationToken).ConfigureAwait(false);
@@ -98,7 +98,7 @@ public sealed class OutboundDispatcher(
         }
         catch (Exception ex)
         {
-            var retryAt = timeProvider.GetUtcNow().UtcDateTime.AddSeconds(_options.BackoffSeconds * (bundle.Attempts + 1));
+            var retryAt = timeProvider.GetUtcNow().UtcDateTime.AddSeconds((double)_options.BackoffSeconds * (bundle.Attempts + 1));
             bundle.MarkAttemptFailed(ex.Message, retryAt, _options.MaxAttempts);
             logger.LogError(ex, "Delivery threw for bundle {Id}", bundle.Id);
             if (bundle.Status == OutboundBundleStatus.Failed)
